fixes `overlay_images` tool `seg_contour` method
The `seg_contour` method of the `overlay_images` tool had two issues. First, it generated outlines whose thickness (and opacity for `thickness<1`) were dependent on the *labels* of the outlined objects. Second, it generated images which were of a different size than the input images. Both issues are resolved by the new implementation, which, however, requires that `thickness` is an integer and `>=1`.